[WIP] Generate JWKs, Save to Disk, Sign, Verify
Building out the bulk of what will be required for our JWT sign/
verify system. We'll have a `JwtToken` struct representing what
our JWT tokens will look like, and we'll have a `JwtIssuer` which
will use the selected private key to sign `JwtTokens`, as well as
a `JwtVerifier` which will use public keys to decode and verify
tokens into `JwtToken` instances.

Once this is out of the way, we can then implement Rocket fairings
to serve as the middleware for verification, and will issue tokens
on successful login back to the client.